requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
alembic
python-dotenv
pyjwt
passlib[bcrypt]